# docs/VMOGD3swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "VMOGD3Project",
    "contact": {}
  },
  "host": "localhost:5000",
  "basePath": "/api/v1",
  "securityDefinitions": {
    "token": {
      "type": "apiKey",
      "description": "",
      "name": "token",
      "in": "header"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/departments": {
      "post": {
        "summary": "Add department",
        "tags": [
          "Department"
        ],
        "operationId": "Adddepartment",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AdddepartmentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "summary": "List departments",
        "tags": [
          "Department"
        ],
        "operationId": "Listdepartments",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/departments/{id}": {
      "get": {
        "summary": "Get department by id",
        "tags": [
          "Department"
        ],
        "operationId": "Getdepartmentbyid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "summary": "Update department",
        "tags": [
          "Department"
        ],
        "operationId": "Updatedepartment",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdatedepartmentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "summary": "Delete department",
        "tags": [
          "Department"
        ],
        "operationId": "Deletedepartment",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AdddepartmentRequest": {
      "title": "AdddepartmentRequest",
      "example": {
        "name": "R&D",
        "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Lectus mauris ultrices eros in cursus turpis massa.",
        "foundingDate": "10/23/2010",
        "adminId": "d00025df-d600-4bef-a2e5-8b1a2eb45404",
        "employees": [
          "f48a8d4c-2b4a-4a1b-913f-c4f9bc8acd83",
          "bb4eb41b-24a9-4e78-b693-7d741261b61b",
          "b5a0e04d-6bcc-4100-b3cc-17a5811882ef"
        ],
        "projects": []
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "foundingDate": {
          "type": "string"
        },
        "adminId": {
          "type": "string"
        },
        "employees": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "projects": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "name",
        "description",
        "foundingDate",
        "adminId",
        "employees",
        "projects"
      ]
    },
    "UpdatedepartmentRequest": {
      "title": "UpdatedepartmentRequest",
      "example": {
        "name": "Research and Development",
        "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Lectus mauris ultrices eros in cursus turpis massa.",
        "foundingDate": "10/23/2010",
        "adminId": "d00025df-d600-4bef-a2e5-8b1a2eb45404",
        "employees": [
          "f48a8d4c-2b4a-4a1b-913f-c4f9bc8acd83",
          "bb4eb41b-24a9-4e78-b693-7d741261b61b",
          "b5a0e04d-6bcc-4100-b3cc-17a5811882ef",
          "8efb0d42-5d65-43ae-83cc-89d6e3bc6d31"
        ],
        "projects": [
          "300dc5e4-a260-41dd-ae79-935f330442bf"
        ]
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "foundingDate": {
          "type": "string"
        },
        "adminId": {
          "type": "string"
        },
        "employees": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "projects": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "name",
        "description",
        "foundingDate",
        "adminId",
        "employees",
        "projects"
      ]
    },
...
  },
  "security": [
    {
      "token": []
    }
  ],
...
}